# faster lsim for cont systems