```
julia> @btime step($sysss, $t);
  689.942 μs (8510 allocations: 3.07 MiB) before

julia> @btime step($sysss, $t);
  632.183 μs (4706 allocations: 2.43 MiB) inplace mul

julia> @btime step($sysss, $t);
  601.745 μs (4706 allocations: 2.43 MiB) inline
```

---

It turns out that the `reduce(hcat, generator)` for `uout` had quadratic time complexity, making it take *far* longer than the solve for long solves.

```julia
julia> @btime step($sysss, $t);
  516.064 μs (4698 allocations: 2.40 MiB) view 
  
julia> @btime step($sysss, $t);
  492.480 μs (4702 allocations: 2.40 MiB) let
  
julia> @btime step($sysss, $t);
  205.648 μs (1674 allocations: 306.44 KiB) loop over uout ( check diff in memory)
```

The problem that led me to find this bug went from
```julia
21.450054 seconds (2.57 M allocations: 116.727 GiB, 31.64% gc time)
```
to
```julia
0.084840 seconds (254.81 k allocations: 99.293 MiB)
```